environmental-service/pyproject.toml
[project]
name = "environmental-service"
version = "0.1.0"
description = "Environmental impact calculation service for PromptPulse using EcoLogits-inspired methodology"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
production = [
    "gunicorn>=21.0.0",
    "prometheus-client>=0.19.0",
]
